Rename ErrorBoundaryPage to ErrorBoundary
