+++ packages/module/patternfly-docs/content/extensions/component-groups/examples/ErrorBoundary/ErrorBoundary.md
@@ -5,28 +5,28 @@ section: extensions
 subsection: Component groups
 # Sidenav secondary level section
 # should be the same for all markdown files
-id: ErrorBoundaryPage
+id: ErrorBoundary
 # Tab (react | react-demos | html | html-demos | design-guidelines | accessibility)
 source: react
 # If you use typescript, the name of the interface to display props for
 # These are found through the sourceProps function provided in patternfly-docs.source.js
-propComponents: ['ErrorBoundaryPage']
+propComponents: ['ErrorBoundary']
 ---
 
-import { ErrorBoundaryPage } from "@patternfly/react-component-groups";
+import { ErrorBoundary } from "@patternfly/react-component-groups";
 
-## Components Usage
+## Component usage
 
 This component reuses the ErrorState component for an app error boundary.
 
-### ErrorBoundaryPage Component
+### ErrorBoundary component
 
-```js file="./ErrorBoundaryPageExample.tsx"
+```js file="./ErrorBoundaryExample.tsx"
 
 ```
 
-### ErrorBoundaryPage Component without error
+### ErrorBoundary component without error
 
-```js file="./ErrorBoundaryPageNoExample.tsx"
+```js file="./ErrorBoundaryNoExample.tsx"
 
 ```
