+++ packages/module/patternfly-docs/content/extensions/component-groups/examples/ErrorBoundary/ErrorBoundaryNoExample.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { ErrorBoundary } from '@patternfly/react-component-groups';
+
+export const BasicExample: React.FunctionComponent = () => (<ErrorBoundary headerTitle={'My app header'}>
+  <div>My app content</div>
+</ErrorBoundary>);
+
