--- packages/module/patternfly-docs/content/extensions/component-groups/examples/ErrorBoundaryPage/ErrorBoundaryPageExample.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { ErrorBoundaryPage } from '@patternfly/react-component-groups';
-
-export const BasicExample: React.FunctionComponent = () => {
-  const Surprise = () => {
-    throw new Error('but a welcome one');
-  };
-
-  return (<ErrorBoundaryPage headerTitle="My app header" errorTitle="Something wrong happened">
-    <Surprise />
-  </ErrorBoundaryPage>);
-}
