--- packages/module/patternfly-docs/content/extensions/component-groups/examples/ErrorBoundaryPage/ErrorBoundaryPageNoExample.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { ErrorBoundaryPage } from '@patternfly/react-component-groups';
-
-export const BasicExample: React.FunctionComponent = () => (<ErrorBoundaryPage headerTitle={'My app header'}>
-  <div>My app content</div>
-</ErrorBoundaryPage>);
-
