+++ packages/module/src/ErrorBoundary/index.ts
@@ -1,0 +1,1 @@
+export { default as ErrorBoundary } from './ErrorBoundary';
