--- packages/module/src/ErrorBoundaryPage/index.ts
@@ -1,1 +1,0 @@
-export { default as ErrorBoundaryPage } from './ErrorBoundaryPage';
