+++ packages/module/src/index.ts
@@ -1,6 +1,6 @@
 export * from './DetailsPage';
 export * from './DetailsPageHeader';
-export * from './ErrorBoundaryPage';
+export * from './ErrorBoundary';
 export * from './ErrorState';
 export * from './HorizontalNav';
 export * from './NotAuthorized';
